Create java.sh
java.sh is a script meant to be run through cron that will check once a day/week/month/etc whether or not the version of java it has on file is up to date. If it's not, it will email someone (passed in as a variable/flag ie: sh java.sh colin@colind.me) letting them know.